requirements.txt
python-dotenv
google-generativeai
chromadb
langchain-text-splitters
tqdm
streamlit